ai/requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
libwapiti==0.2.1
MarkupSafe==2.1.5
nltk==3.3
pymongo==4.7.3
python-dotenv==1.0.1
Werkzeug==3.0.3
